Doubles and fixes
Tech and Energy are now doubles - had to be done - and a number of fixes
are in here.

+++ AlienInterfaces/src/alieninterfaces/Context.java
@@ -11,9 +11,9 @@ package alieninterfaces;
  */
 public interface Context {
 
-    int getEnergy();
+    double getEnergy();
 
-    int getTech();
+    double getTech();
 
     int getX();
