+++ GameEngine/src/gameengineinterfaces/AlienSpec.java
@@ -20,14 +20,14 @@ public class AlienSpec {
     public int hashCode;
     public int x;
     public int y;
-    public int tech;
-    public int energy;
-    public int actionPower;
+    public double tech;
+    public double energy;
+    public double actionPower;
     public String fullName;
     public String speciesName;
 
     public AlienSpec(String domainName, String packageName, String className,
-            int speciesID, int hashCode, int x, int y, int tech, int energy,
+            int speciesID, int hashCode, int x, int y, double tech, double energy,
             String fullName, String speciesName) {
         this.domainName = domainName;
         this.packageName = packageName;
@@ -43,8 +43,8 @@ public class AlienSpec {
     }
 
     public AlienSpec(String domainName, String packageName, String className,
-            int speciesID, int hashCode, int x, int y, int tech, int energy,
-            String fullName, String speciesName, int actionPower) {
+            int speciesID, int hashCode, int x, int y, double tech, double energy,
+            String fullName, String speciesName, double actionPower) {
         this.domainName = domainName;
         this.packageName = packageName;
         this.className = className;
