+++ StockAliens/src/stockaliens/Dalek.java
@@ -6,7 +6,6 @@
 package stockaliens;
 
 import alieninterfaces.*;
-import java.util.Random;
 
 /**
  *
@@ -26,8 +25,8 @@ public class Dalek implements Alien {
         ctx.debugOut("Initialized at "
                 + "(" + Integer.toString(ctx.getX())
                 + "," + Integer.toString(ctx.getY()) + ")"
-                + " E: " + Integer.toString(ctx.getEnergy())
-                + " T: " + Integer.toString(ctx.getTech()));
+                + " E: " + Double.toString(ctx.getEnergy())
+                + " T: " + Double.toString(ctx.getTech()));
 
     }
 
@@ -35,15 +34,15 @@ public class Dalek implements Alien {
     public MoveDir getMove() {
 
         ctx.debugOut("Move requested,"
-                + " E:" + Integer.toString(ctx.getEnergy())
-                + " T:" + Integer.toString(ctx.getTech()));
+                + " E:" + Double.toString(ctx.getEnergy())
+                + " T:" + Double.toString(ctx.getTech()));
         int move_energy;
 
         if (ctx.getEnergy() < 30) {
             return new MoveDir(0,0);
         }
         // don't move more than you have tech
-        move_energy = Math.min(ctx.getTech(), ctx.getEnergy());
+        move_energy = (int)Math.min(ctx.getTech(), ctx.getEnergy());
         // don't move more than 5, leave energy for other stuff
         move_energy = Math.min(move_energy, 10);
 
@@ -61,26 +60,26 @@ public class Dalek implements Alien {
     public Action getAction() {
 
         ctx.debugOut("Action requested,"
-                + " E:" + Integer.toString(ctx.getEnergy())
-                + " T:" + Integer.toString(ctx.getTech()));
+                + " E:" + Double.toString(ctx.getEnergy())
+                + " T:" + Double.toString(ctx.getTech()));
         View view = ctx.getView();
 
         // catch any shenanigans
         try {
             if (view.getAlienCountAtPos(ctx.getX(), ctx.getY()) > 1) {
                 ctx.debugOut("Uh-oh.There is someone else here."
-                        + " E:" + Integer.toString(ctx.getEnergy())
-                        + " T:" + Integer.toString(ctx.getTech())
-                        + " X:" + Integer.toString(ctx.getX())
-                        + " Y:" + Integer.toString(ctx.getY()));
+                        + " E:" + Double.toString(ctx.getEnergy())
+                        + " T:" + Double.toString(ctx.getTech())
+                        + " X:" + Double.toString(ctx.getX())
+                        + " Y:" + Double.toString(ctx.getY()));
             }
 
             // do we have enough energy?
             if (ctx.getEnergy() < 10) {
                 // no, charge
                 ctx.debugOut("Choosing to gain energy,"
-                        + " E:" + Integer.toString(ctx.getEnergy())
-                        + " T:" + Integer.toString(ctx.getTech()));
+                        + " E:" + Double.toString(ctx.getEnergy())
+                        + " T:" + Double.toString(ctx.getTech()));
                 return new Action(ActionCode.Gain);
             }
 
@@ -88,27 +87,27 @@ public class Dalek implements Alien {
             if (ctx.getTech() < 30 && ctx.getEnergy() > ctx.getTech()) {
                 // no, research
                 ctx.debugOut("Choosing to research"
-                        + " E:" + Integer.toString(ctx.getEnergy())
-                        + " T:" + Integer.toString(ctx.getTech()));
+                        + " E:" + Double.toString(ctx.getEnergy())
+                        + " T:" + Double.toString(ctx.getTech()));
                 return new Action(ActionCode.Research);
             }
 
             // is there another alien on our position?
             if (view.getAlienCountAtPos(ctx.getX(), ctx.getY()) > 1) {
                 ctx.debugOut("EXTERMINATE!!!!!"
-                        + " E:" + Integer.toString(ctx.getEnergy())
-                        + " T:" + Integer.toString(ctx.getTech())
-                        + " X:" + Integer.toString(ctx.getX())
-                        + " Y:" + Integer.toString(ctx.getY()));
+                        + " E:" + Double.toString(ctx.getEnergy())
+                        + " T:" + Double.toString(ctx.getTech())
+                        + " X:" + Double.toString(ctx.getX())
+                        + " Y:" + Double.toString(ctx.getY()));
 
-                return new Action(ActionCode.Fight, ctx.getEnergy() - 2);
+                return new Action(ActionCode.Fight, (int)ctx.getEnergy() - 2);
             }
 
             if (ctx.getEnergy() > (ctx.getSpawningCost() + 10)) {
                 // no other aliens here, have enough stuff, spawn!
                 ctx.debugOut("DALEKS RULE SUPREME! SPAWNING!"
-                        + " E:" + Integer.toString(ctx.getEnergy())
-                        + " T:" + Integer.toString(ctx.getTech()));
+                        + " E:" + Double.toString(ctx.getEnergy())
+                        + " T:" + Double.toString(ctx.getTech()));
 
                 return new Action(ActionCode.Spawn, 5);
             }
